feat: add exchange config with rates and fee (Step 2)
Introduces config/exchange.php with rate keys in {from}_to_{to} format
and a fee_percent applied to the gross credited amount (not the source).

- gold_to_gems = 0.1, gems_to_gold = 8.0, fee_percent = 2.5
- ExchangeService now reads rates from config and applies fee:
  net = amount * rate * (1 - fee_percent/100)
  e.g. 100 gold -> 10 gems gross -> 9.75 gems net
- balance and to_amount columns changed to decimal(20,8) since
  credited amounts are fractional after rate+fee arithmetic
- float cast added to Wallet.balance and ExchangeLog.to_amount
- All 20 tests updated and passing (feature + concurrency)

+++ database/migrations/2026_06_04_000001_create_wallets_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->enum('currency', ['gold', 'gems']);
-            $table->unsignedBigInteger('balance')->default(0);
+            $table->decimal('balance', 20, 8)->default(0);
             $table->timestamps();
 
             $table->unique(['user_id', 'currency']);
